chore: remove unreachable code
