--- transaction_test.go
@@ -1,288 +1,0 @@
-package sqld
-
-import (
-	"context"
-	"database/sql"
-	"errors"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
-)
-
-// Mock implementations for testing
-
-type MockDB struct {
-	mock.Mock
-}
-
-func (m *MockDB) Query(ctx context.Context, query string, args ...interface{}) (Rows, error) {
-	mockArgs := append([]interface{}{ctx, query}, args...)
-	ret := m.Called(mockArgs...)
-	return ret.Get(0).(Rows), ret.Error(1)
-}
-
-func (m *MockDB) QueryRow(ctx context.Context, query string, args ...interface{}) Row {
-	mockArgs := append([]interface{}{ctx, query}, args...)
-	ret := m.Called(mockArgs...)
-	return ret.Get(0).(Row)
-}
-
-func (m *MockDB) BeginTx(ctx context.Context, opts *TxOptions) (Tx, error) {
-	ret := m.Called(ctx, opts)
-	return ret.Get(0).(Tx), ret.Error(1)
-}
-
-func (m *MockDB) WithTransaction(ctx context.Context, opts *TxOptions, fn func(ctx context.Context, tx Tx) error) error {
-	ret := m.Called(ctx, opts, fn)
-	return ret.Error(0)
-}
-
-type MockTx struct {
-	mock.Mock
-}
-
-func (m *MockTx) Query(ctx context.Context, query string, args ...interface{}) (Rows, error) {
-	mockArgs := append([]interface{}{ctx, query}, args...)
-	ret := m.Called(mockArgs...)
-	return ret.Get(0).(Rows), ret.Error(1)
-}
-
-func (m *MockTx) QueryRow(ctx context.Context, query string, args ...interface{}) Row {
-	mockArgs := append([]interface{}{ctx, query}, args...)
-	ret := m.Called(mockArgs...)
-	return ret.Get(0).(Row)
-}
-
-func (m *MockTx) Commit(ctx context.Context) error {
-	ret := m.Called(ctx)
-	return ret.Error(0)
-}
-
-func (m *MockTx) Rollback(ctx context.Context) error {
-	ret := m.Called(ctx)
-	return ret.Error(0)
-}
-
-func (m *MockTx) Exec(ctx context.Context, query string, args ...interface{}) (sql.Result, error) {
-	mockArgs := append([]interface{}{ctx, query}, args...)
-	ret := m.Called(mockArgs...)
-	return ret.Get(0).(sql.Result), ret.Error(1)
-}
-
-type MockRows struct {
-	mock.Mock
-}
-
-func (m *MockRows) Close() error {
-	ret := m.Called()
-	return ret.Error(0)
-}
-
-func (m *MockRows) Next() bool {
-	ret := m.Called()
-	return ret.Bool(0)
-}
-
-func (m *MockRows) Scan(dest ...interface{}) error {
-	ret := m.Called(dest...)
-	return ret.Error(0)
-}
-
-func (m *MockRows) Err() error {
-	ret := m.Called()
-	return ret.Error(0)
-}
-
-type MockRow struct {
-	mock.Mock
-}
-
-func (m *MockRow) Scan(dest ...interface{}) error {
-	ret := m.Called(dest...)
-	return ret.Error(0)
-}
-
-type MockResult struct {
-	mock.Mock
-}
-
-func (m *MockResult) LastInsertId() (int64, error) {
-	ret := m.Called()
-	return ret.Get(0).(int64), ret.Error(1)
-}
-
-func (m *MockResult) RowsAffected() (int64, error) {
-	ret := m.Called()
-	return ret.Get(0).(int64), ret.Error(1)
-}
-
-func TestStandardTx_Query(t *testing.T) {
-	t.Run("query validation error", func(t *testing.T) {
-		// Create a dummy sql.Tx - we'll skip testing actual DB operations
-		tx := NewStandardTx(nil, Postgres)
-
-		ctx := context.Background()
-		query := "" // Invalid empty query
-
-		_, err := tx.Query(ctx, query)
-		assert.Error(t, err)
-
-		var qErr *QueryError
-		assert.True(t, errors.As(err, &qErr))
-
-		var vErr *ValidationError
-		assert.True(t, errors.As(qErr.Unwrap(), &vErr))
-	})
-}
-
-func TestStandardTx_QueryRow(t *testing.T) {
-	t.Run("query validation error", func(t *testing.T) {
-		tx := NewStandardTx(nil, Postgres)
-
-		ctx := context.Background()
-		query := "" // Invalid empty query
-
-		row := tx.QueryRow(ctx, query)
-
-		// Should return ErrorRow
-		errorRow, ok := row.(*ErrorRow)
-		assert.True(t, ok)
-		assert.Error(t, errorRow.err)
-
-		var qErr *QueryError
-		assert.True(t, errors.As(errorRow.err, &qErr))
-
-		var vErr *ValidationError
-		assert.True(t, errors.As(qErr.Unwrap(), &vErr))
-	})
-}
-
-func TestTransactionalQueries_WithTx(t *testing.T) {
-	t.Skip("Skipping complex transaction test - would require full mock setup")
-
-	// This test would require a complete implementation of mock transaction manager
-	// For now, we'll test the structure and basic error handling elsewhere
-}
-
-func TestRunInTransaction(t *testing.T) {
-	mockTxManager := &MockDB{}
-
-	ctx := context.Background()
-
-	// Test with successful operations
-	op1 := func(ctx context.Context, tx Tx) error {
-		return nil
-	}
-
-	op2 := func(ctx context.Context, tx Tx) error {
-		return nil
-	}
-
-	mockTxManager.On("WithTransaction", ctx, (*TxOptions)(nil), mock.AnythingOfType("func(context.Context, sqld.Tx) error")).Return(nil)
-
-	err := RunInTransaction(ctx, mockTxManager, nil, op1, op2)
-	assert.NoError(t, err)
-
-	mockTxManager.AssertExpectations(t)
-}
-
-func TestErrorRow(t *testing.T) {
-	testErr := errors.New("test error")
-	errorRow := &ErrorRow{err: testErr}
-
-	var dest string
-	err := errorRow.Scan(&dest)
-
-	assert.Equal(t, testErr, err)
-}
-
-func TestTxOptions(t *testing.T) {
-	opts := &TxOptions{
-		IsolationLevel: sql.LevelReadCommitted,
-		ReadOnly:       true,
-	}
-
-	assert.Equal(t, sql.LevelReadCommitted, opts.IsolationLevel)
-	assert.True(t, opts.ReadOnly)
-}
-
-// Benchmark tests for performance
-func BenchmarkValidateQuery(b *testing.B) {
-	query := "SELECT * FROM users WHERE name = $1 AND status = $2"
-
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		ValidateQuery(query, Postgres)
-	}
-}
-
-func BenchmarkWhereBuilderSimple(b *testing.B) {
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		where := NewWhereBuilder(Postgres)
-		where.Equal("name", "John")
-		where.Equal("status", "active")
-		where.Build()
-	}
-}
-
-func BenchmarkWhereBuilderComplex(b *testing.B) {
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		where := NewWhereBuilder(Postgres)
-		where.Equal("name", "John")
-		where.GreaterThan("age", 18)
-		where.In("role", []interface{}{"admin", "user", "manager"})
-		where.Or(func(or ConditionBuilder) {
-			or.Like("email", "%@company.com")
-			or.IsNull("deleted_at")
-		})
-		where.Build()
-	}
-}
-
-// Integration test helpers (would require actual database)
-func TestIntegrationHelper(t *testing.T) {
-	t.Skip("Skipping integration test - requires actual database")
-
-	// Example of how integration tests would be structured:
-	// 1. Setup test database
-	// 2. Create StandardDB instance
-	// 3. Test actual transaction operations
-	// 4. Verify data consistency
-	// 5. Cleanup
-}
-
-// Test timeout behavior
-func TestContextTimeout(t *testing.T) {
-	t.Run("context timeout during validation", func(t *testing.T) {
-		// Create a context that's already cancelled
-		ctx, cancel := context.WithCancel(context.Background())
-		cancel()
-
-		// The validation itself doesn't check context, but it's good practice
-		// to pass context through for future enhancements
-		query := "SELECT * FROM users"
-		err := ValidateQuery(query, Postgres)
-
-		// Validation should still work even with cancelled context
-		assert.NoError(t, err)
-
-		// But context should be cancelled
-		assert.Error(t, ctx.Err())
-	})
-
-	t.Run("context timeout simulation", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 1*time.Millisecond)
-		defer cancel()
-
-		// Wait for timeout
-		time.Sleep(2 * time.Millisecond)
-
-		// Context should be cancelled
-		assert.Error(t, ctx.Err())
-		assert.Equal(t, context.DeadlineExceeded, ctx.Err())
-	})
-}
